Add tt-installer to Quick Installation step
Now that tt-installer is mature enough to be used
in production, we can add this Quick Installation step to the Quickstart
for users who feel comfortable with a basic installation. For users who
want a more customized installation or who feel uncomfortable using the
script, we should leave the existing instructions under the header of
Manual Installation.

Signed-off-by: June Knauth (tt) <jknauth@tenstorrent.com>